% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPR190512.R
\name{d2rsq.2}
\alias{d2rsq.2}
\title{New implmentation of \code{d2rsq}}
\usage{
d2rsq.2(d, p = NULL, var.names = NULL, expand = FALSE,
  return.df = FALSE)
}
\arguments{
\item{d}{Either a \code{data.frame} which contains columns corresponding to
Cohen's d and the proportion, or a numeric vector of Cohen's d values.}

\item{p}{Numeric vector of proportions. The defualt value, \code{NULL},
should be used if \code{d} is a \code{data.frame}.}

\item{var.names}{Ordered character vector with column names of columns
corresponding to Cohen's d and the proportion. The name of the column
corresponding to Cohen's d must come first \code{var.names} only needs to
be specified if the columne names in the \code{data.frame} differ from "d"
and "p".}

\item{expand}{logical. If \code{TRUE} \code{expand.grid} is used to create
all combinations of the provided Cohen's d values and proportions.}

\item{return.df}{logical. If \code{FALSE} the function returns a vector of R
squared values. If \code{TRUE} the function returns a \code{data.frame}
containing all Cohen's d values and proportions with corresponding R
squared values.}
}
\description{
Adds parameter checks and expands features
}
