% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPR190512.R
\name{d2rsq}
\alias{d2rsq}
\title{Get Rsquare from Cohen's d}
\usage{
d2rsq(d, p)
}
\arguments{
\item{d}{not sure}

\item{p}{not sure

Detailed explanation of what the function does and how}
}
\description{
Slightly longer description here
}
