% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPR190512.R
\name{power.lm.fn}
\alias{power.lm.fn}
\title{Compute power in linear model}
\usage{
power.lm.fn(n, alpha, df.test, df.adj, r2.full, r2.red)
}
\arguments{
\item{n}{sample size}

\item{alpha}{significance level (type 1 error rate)}

\item{df.test}{number of degrees of freedom for test}

\item{df.adj}{adjusted degrees of freedom}

\item{r2.full}{r squared full}

\item{r2.red}{r squared reduced}
}
\value{
Compute power in linear model (more description)
}
\description{
Slightly longer description here
}
\examples{
power.lm.fn(n = 180, alpha = 0.05, df.test = 2, df.adj = 4 + 1, r2.full = 0.1, r2.red = 0)

}
