% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPR190512.R
\name{sim.design.fn}
\alias{sim.design.fn}
\title{Design Constructor (breif description goes here)}
\usage{
sim.design.fn(x)
}
\arguments{
\item{x}{a list of "factors" used in simulation design}
}
\value{
A longer description of what the function does can go here.
}
\description{
You can add a slightly longer description here which is displayed underneath
the brief description
}
