% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPR190512.R
\name{power.lm.fn}
\alias{power.lm.fn}
\title{Compute power in linear model}
\usage{
power.lm.fn(n, alpha, dfTest, dfAdj, R2full, R2red)
}
\arguments{
\item{n}{sample size}

\item{alpha}{significance level (type 1 error rate)}

\item{dfTest}{number of degrees of freedom for test}

\item{dfAdj}{adjusted degrees of freedom}

\item{R2full}{r squared full}

\item{R2red}{r squared reduced}
}
\value{
Compute power in linear model (more description)SSDFDS SDFO SDFO
  IDIOJSD FSKD SODIFJ SOIJS DOFI OSDIFJS D
}
\description{
Compute power in linear model
}
\examples{
power.lm.fn(n = 180, alpha = 0.05, dfTest = 2, dfAdj = 4 + 1, R2full = 0.1, R2red = 0)

}
